+++ src/main/java/com/onlyoffice/model/convertservice/ConvertResponse.java
@@ -113,7 +113,12 @@ public class ConvertResponse {
         /**
          * Invalid token (the error code is -8).
          */
-        TOKEN(-8, "Invalid token");
+        TOKEN(-8, "Invalid token"),
+
+        /**
+         * Error automatically determine the output file format (the error code is -9).
+         */
+        OOXML_OUTPUT_TYPE(-9, "Error automatically determine the output file format");
 
         /**
          * Defines the error code.